+++ tests/macro_read/enum_without_id_type.rs
@@ -1,6 +1,0 @@
-use deku::prelude::*;
-
-#[derive(DekuRead)]
-enum Test {}
-
-fn main() {}